fix: поправил замечания к Diplom_2
