+++ src/main/java/praktikum/order/IngredientsGenerator.java
@@ -1,4 +1,4 @@
-package praktikum.Order;
+package praktikum.order;
 
 public class IngredientsGenerator {
