+++ src/main/java/praktikum/user/UserGenerator.java
@@ -1,0 +1,29 @@
+package praktikum.user;
+
+import org.apache.commons.lang3.RandomStringUtils;
+
+public class UserGenerator {
+
+    private static int count = 10;
+    private static int countIncorrectPassword = 2;
+
+    public static String generateString(int count) {
+        return RandomStringUtils.randomAlphabetic(count);
+    }
+
+    public static User getUser() {
+        return new User(generateString(count) + "@mail.ru", generateString(count), generateString(count));
+    }
+
+    public static User getUserIncorrect() {
+        return new User(generateString(count) + "@mail.ru", generateString(countIncorrectPassword), generateString(count));
+    }
+
+    public static User getUserUpdate() {
+        return new User(generateString(count) + "@mail.ru", generateString(count), generateString(count));
+    }
+
+    public static User getUserSecond() {
+        return new User(generateString(count) + "@mail.ru", generateString(count), generateString(count));
+    }
+}
