+++ src/test/java/user/UserCreateTest.java
@@ -4,10 +4,10 @@ import io.restassured.response.ValidatableResponse;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import praktikum.User.Credentials;
-import praktikum.User.User;
-import praktikum.User.UserGenerator;
-import praktikum.User.UserClient;
+import praktikum.user.Credentials;
+import praktikum.user.User;
+import praktikum.user.UserGenerator;
+import praktikum.user.UserClient;
 
 import static org.apache.http.HttpStatus.*;
 import static org.junit.Assert.assertEquals;
@@ -17,11 +17,13 @@ public class UserCreateTest {
     private User user;
     private UserClient userClient;
     private String accessToken;
+    private String accessTokenError;
 
     @Before
     public void setUp() {
         user = UserGenerator.getUser();
         userClient = new UserClient();
+        accessTokenError = null;
     }
 
     @Test
@@ -36,6 +38,9 @@ public class UserCreateTest {
         ValidatableResponse responseCreateUniqueUser = userClient.createUser(Credentials.from(user));
         ValidatableResponse responseCreateExistUser = userClient.createUser(Credentials.from(user));
         accessToken = responseCreateUniqueUser.extract().path("accessToken");
+        if (responseCreateExistUser.extract().path("accessToken") != null) {
+            accessTokenError = responseCreateExistUser.extract().path("accessToken");
+        }
         assertEquals(SC_FORBIDDEN, responseCreateExistUser.extract().statusCode());
         assertEquals("User already exists", responseCreateExistUser.extract().path("message"));
     }
@@ -43,6 +48,9 @@ public class UserCreateTest {
     @Test
     public void createUserWithoutNameAndPassword403(){
         ValidatableResponse responseCreate = userClient.createUser(Credentials.fromOnlyEmail(user));
+        if (responseCreate.extract().path("accessToken") != null) {
+            accessTokenError = responseCreate.extract().path("accessToken");
+        }
         assertEquals(SC_FORBIDDEN, responseCreate.extract().statusCode());
         assertEquals("Email, password and name are required fields", responseCreate.extract().path("message"));
     }
@@ -50,6 +58,9 @@ public class UserCreateTest {
     @Test
     public void createUserWithoutNameAndEmail403(){
         ValidatableResponse responseCreate = userClient.createUser(Credentials.fromOnlyPassword(user));
+        if (responseCreate.extract().path("accessToken") != null) {
+            accessTokenError = responseCreate.extract().path("accessToken");
+        }
         assertEquals(SC_FORBIDDEN, responseCreate.extract().statusCode());
         assertEquals("Email, password and name are required fields", responseCreate.extract().path("message"));
     }
@@ -57,14 +68,18 @@ public class UserCreateTest {
     @Test
     public void createUserWithoutName403(){
         ValidatableResponse responseCreate = userClient.createUser(Credentials.fromOnlyEmailAndPassword(user));
+        if (responseCreate.extract().path("accessToken") != null) {
+            accessTokenError = responseCreate.extract().path("accessToken");
+        }
         assertEquals(SC_FORBIDDEN, responseCreate.extract().statusCode());
         assertEquals("Email, password and name are required fields", responseCreate.extract().path("message"));
     }
 
-
-
     @After
     public void cleanUp() {
         userClient.deleteUser(accessToken);
+        if (accessTokenError != null) {
+            userClient.deleteUser(accessTokenError);
+        }
     }
 }
